+++ src/Aresta.cpp
@@ -1,6 +1,6 @@
 #include "Aresta.h"
 
-Aresta::Aresta(int id_No, double peso) {
+Aresta::Aresta(int id_No, int peso) {
     this->pesoAresta    = peso;
     this->id_No         = id_No;
 }
@@ -17,6 +17,6 @@ int Aresta::getIDNo() {
     return this->id_No;
 }
 
-void  Aresta::setPesoAresta(double peso) {
+void  Aresta::setPesoAresta(int peso) {
     this->pesoAresta = peso;
 }
